refactor: ゲームに関するオブジェクトの分離
