+++ js/gameObject/Obstacle.mjs
@@ -1,0 +1,9 @@
+// 障害物の親クラス
+export class Obstacle {
+    constructor(x, d) {
+        this.type = null;
+        this.x = x;
+        this.d = d;
+        this.image = null;
+    }
+}
