+++ js/scene/DriveScene.mjs
@@ -1,124 +1,8 @@
 import { Scene } from './special/Scene.mjs';
 import { scenes } from "./special/sceneSettings.mjs";
 import { stage1 } from '../stage/stage1.mjs';
-import { obstacleType } from '../stage/enum.mjs';
-
-class Player {
-    constructor(x) {
-        this.x = x;
-        this.d = 0;
-        this.xControlSpeed = 30; // px/s
-        this.dControlSpeed = 30; // px/s
-        this.dSpeed = 60; // px/s
-        this.image = new Image();
-        this.image.src = '../../../resource/image/rider.png';
-        this.theta = 0;
-        this.inCollision = false;
-    }
-
-    draw(max_x, max_y, ctx, pixelSize, cameraDistance) {
-        const y = max_y - (this.d - cameraDistance) * pixelSize;
-        const x = this.x * pixelSize;
-        const scaleFactor = 1.5;
-        const centerX = x - this.image.width * scaleFactor / 2;
-        const centerY = y - this.image.height * scaleFactor / 2;
-
-        if (this.image.complete) {
-            ctx.save();
-            ctx.imageSmoothingEnabled = false;
-            ctx.translate(
-                centerX + this.image.width * scaleFactor / 2,
-                centerY + this.image.height * scaleFactor / 2
-            );
-            ctx.rotate(this.theta);
-            ctx.drawImage(
-                this.image,
-                -this.image.width * scaleFactor / 2,
-                -this.image.height * scaleFactor / 2,
-                this.image.width * scaleFactor,
-                this.image.height * scaleFactor
-            );
-            ctx.restore();
-        }
-    }
-
-    updatePosition(deltaTime, leftPressed, rightPressed, upPressed, downPressed) {
-        if (this.inCollision) return;
-        this.d += this.dSpeed * deltaTime / 1000
-        if (leftPressed) {
-            this.x -= this.xControlSpeed * deltaTime / 1000
-        }
-        if (rightPressed) {
-            this.x += this.xControlSpeed * deltaTime / 1000
-        }
-        if (upPressed) {
-            this.d += this.dControlSpeed * deltaTime / 1000
-        }
-        if (downPressed) {
-            this.d -= this.dControlSpeed * deltaTime / 1000
-        }
-        const currentXSpeed = ((rightPressed - leftPressed) * this.xControlSpeed);
-        const currentDSpeed = ((upPressed - downPressed) * this.dControlSpeed) + this.dSpeed;
-        this.theta = Math.atan(currentXSpeed / currentDSpeed);
-    }
-}
-
-// 障害物の親クラス
-class Obstacle {
-    constructor(x, d) {
-        this.type = null;
-        this.x = x;
-        this.d = d;
-        this.image = null;
-    }
-}
-
-function makeObstacle(type, x, d) {
-    switch (type) {
-        case obstacleType.mud:
-            return new Mud(x, d);
-        default:
-            console.error(`makeObstacleに未定義のtypeが渡されました：${type}`)
-    }
-}
-
-class Mud extends Obstacle {
-    constructor(x, d) {
-        super(x, d);
-        this.type = obstacleType.mud;
-        this.image = new Image();
-        this.image.src = '../../../resource/image/mud.png';
-    }
-
-    draw(max_x, max_y, ctx, pixelSize, cameraDistance) {
-        const y = max_y - (this.d - cameraDistance) * pixelSize;
-        const x = this.x * pixelSize;
-        const scaleFactor = 1.5;
-
-        if (this.image.complete) {
-            ctx.imageSmoothingEnabled = false;
-            ctx.drawImage(
-                this.image,
-                x - this.image.width * scaleFactor / 2,
-                y - this.image.height * scaleFactor / 2,
-                this.image.width * scaleFactor,
-                this.image.height * scaleFactor,
-            );
-        }
-    }
-
-    checkCollision(x, d) {
-        return (this.x - x)**2 + (this.d - d)**2 <= 3**2;
-    }
-
-    handleCollision(player, roadX) {
-        player.inCollision = true;
-        setTimeout(() => {
-            player.inCollision = false;
-            player.x = roadX.center;
-        }, 1000);
-    }
-}
+import { obstacleType, makeObstacle } from '../gameObject/obstacleSettings.mjs';
+import { Player } from '../gameObject/Player.mjs';
 
 export class DriveScene extends Scene {
     sceneWillAppear() {
