--- js/stage/enum.mjs
@@ -1,3 +1,0 @@
-export const obstacleType = {
-    mud: "obstacleType.mud"
-}
